Move balance and order into own file
without any other logic change
