+++ driver/src/orderbook/streamed/state.rs
@@ -4,8 +4,9 @@ use crate::contracts::{
     stablex_contract::batch_exchange::{event_data::*, Event},
 };
 use crate::models::Order as ModelOrder;
+use balance::{Balance, Flux};
 use error::Error;
-use ethcontract::U256;
+use order::Order;
 use serde::{Deserialize, Serialize};
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
@@ -184,84 +185,6 @@ impl Tokens {
             .iter()
             .find(|(_, address_)| *address_ == address)
             .map(|(token_id, _)| *token_id)
-    }
-}
-
-#[derive(Copy, Clone, Debug, Deserialize, Serialize)]
-struct Order {
-    buy_token: TokenId,
-    sell_token: TokenId,
-    valid_from: BatchId,
-    valid_until: BatchId,
-    price_numerator: u128,
-    price_denominator: u128,
-    used_amount: u128,
-}
-
-impl Order {
-    fn has_limited_amount(&self) -> bool {
-        self.price_numerator != std::u128::MAX && self.price_denominator != std::u128::MAX
-    }
-}
-
-#[derive(Copy, Clone, Debug, Deserialize, Serialize)]
-struct Flux {
-    amount: U256,
-    batch_id: BatchId,
-}
-
-#[derive(Clone, Copy, Debug, Default, Deserialize, Serialize)]
-struct Balance {
-    balance: U256,
-    pending_deposit: Option<Flux>,
-    pending_withdraw: Option<Flux>,
-}
-
-impl Balance {
-    fn deposit(&mut self, new_deposit: Flux, current_batch_id: BatchId) {
-        self.update_deposit_balance(current_batch_id);
-        match self.pending_deposit.as_mut() {
-            Some(deposit) => {
-                deposit.amount += new_deposit.amount;
-                deposit.batch_id = new_deposit.batch_id
-            }
-            None => self.pending_deposit = Some(new_deposit),
-        }
-    }
-
-    fn withdraw(&mut self, amount: U256, current_batch_id: BatchId) -> Result<(), Error> {
-        self.update_deposit_balance(current_batch_id);
-        self.pending_withdraw = None;
-        self.balance = self
-            .balance
-            .checked_sub(amount)
-            .ok_or(Error::MathUnderflow)?;
-        Ok(())
-    }
-
-    fn update_deposit_balance(&mut self, current_batch_id: BatchId) {
-        match self.pending_deposit {
-            Some(ref deposit) if deposit.batch_id < current_batch_id => {
-                self.balance += deposit.amount;
-                self.pending_deposit = None;
-            }
-            _ => (),
-        };
-    }
-
-    fn current_balance(&self, current_batch_id: BatchId) -> U256 {
-        let mut balance = self.balance;
-        if let Some(ref flux) = self.pending_deposit {
-            if flux.batch_id < current_batch_id {
-                balance = balance.saturating_add(flux.amount);
-            }
-        }
-        if let Some(ref flux) = self.pending_withdraw {
-            if flux.batch_id < current_batch_id {
-                balance = balance.saturating_sub(flux.amount);
-            }
-        }
-        balance
     }
 }
